moved states from BoostModal  to BoostInnerModal
